test(sys): add stmt2 test case
